requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
isort==6.0.0
Mako==1.3.9
MarkupSafe==3.0.2
oauth2==1.9.0.post1
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyparsing==3.2.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
